Add remaining projects
Added the remaining 3 project objects to the list. Modified the createProjectCard function to check if links are empty, because the last project only has a github link.
Adds a class of hidden to the anchor tag when the link is empty.

Set a bottom margin on the project list of 70px to give more scrolling space at the end.